sync-zabbix-config/requirements.txt
argparse
py-zabbix

# Lint
flake8

# Test
pytest
pytest-html

# Docs
Sphinx
